NIFI-10585: Incorporated additional review comments
