--- nifi-nar-bundles/nifi-standard-bundle/nifi-standard-processors/src/test/resources/TestGenerateFakeRecord/nested_no_nullable.avsc
@@ -1,52 +1,0 @@
-{
-	"type": "record",
-	"name": "nifiRecord",
-	"namespace": "org.apache.nifi",
-	"fields": [{
-		"name": "System",
-		"type": {
-			"type": "record",
-			"name": "SystemType",
-			"fields": [{
-				"name": "Provider",
-				"type": {
-					"type": "record",
-					"name": "ProviderType",
-					"fields": [{
-						"name": "Guid",
-						"type": "string"
-					}, {
-						"name": "Name",
-						"type": "string"
-					}]
-				}
-			}, {
-				"name": "EventID",
-				"type": "int"
-			}]
-		}
-	},{
-		"name": "EventData",
-		"type": {
-			"type": "record",
-			"name": "EventDataType",
-			"fields": [{
-				"name": "Data",
-				"type": {
-					"type": "array",
-					"items": {
-						"type": "record",
-						"name": "DataElementRecord",
-						"fields": [{
-							"name": "Name",
-							"type": "string"
-						}, {
-							"name": "DataElement",
-							"type": "string"
-						}]
-					}
-				}
-			}]
-		}
-	}]
-}Z
